Complete rewrite - monitor overlays show
